+++ changelog.md
@@ -11,6 +11,7 @@
 - Add type menu title to type listing for better sparation
 - Allow BelongsTo field to be extended 
 - Allow operator to be changed for `whereValue` entity scope
+- Change EntityPath fillable to guarded
 
 ## Version 1.4.1 (2017-03-14)
 - Remove debug